+++ src/main/java/com/example/producttestapi/repos/VoucherRepo.java
@@ -3,6 +3,6 @@ package com.example.producttestapi.repos;
 import com.example.producttestapi.entities.Voucher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VoucherRepo extends JpaRepository<Voucher, Long> {
+public interface VoucherRepo extends JpaRepository<Voucher, Integer> {
     Voucher findByCode(String code);
 }
